sales module

+++ app/Http/Livewire/Admin/users/User.php
@@ -5,53 +5,47 @@ namespace App\http\Livewire\Admin\users;
 use App\Models\Admin;
 use App\Models\Company;
 use Livewire\Component;
-use Livewire\WithPagination;
-use Livewire\WithFileUploads;
-use Illuminate\Http\UploadedFile;
+use App\Traits\HasTable;
+use App\services\UserService;
+use App\Traits\HasMultiSelect;
+use App\Traits\HasPhotoUpload;
+use Livewire\Attributes\Computed;
 use Spatie\Permission\Models\Role;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
-
-
-use Illuminate\Support\Facades\Storage;
-
 
 class User extends Component
 {
-    use WithFileUploads;
-    use WithPagination;
-    protected $paginationTheme = 'bootstrap';
 
+    use HasTable;
+    use HasPhotoUpload;
+    use HasMultiSelect;
 
-    public $search;
-    public $perpage =5;
-    public $sortfilter ='desc';
     public $admin_id;
     public $name;
     public $company_id;
     public $email;
     public $password;
     public $phone = null;
-    public $photo;
     public $assigned_roles =[];
-    public $user;
+    public $roles;
+    public $companies;
+    protected $write_permission='write user';
+
+
+     
+    public function mount(){
+        $this->check_permission('view user');
+        $this->roles = Role ::all('id','name');
+        $this->companies = Company::all('id','name');
+    }
+
+    public function closeModal()
+    {
+   
+        $this->reset('name','email','company_id','phone','password','assigned_roles','photo',);
+    }
     
 
-
-
-    public function mount()
-    {
-       $this->user= Auth::user()->name; 
-    }
-
-    public function hydrate(){
-       $this->dispatchBrowserEvent('pharaonic.select2.init');
-       $this->dispatchBrowserEvent('pharaonic.select2.load', [
-           'component' => $this->id,
-           'target'    => '#multiSelect'
-       ]);
-    }
-
+ 
 
     protected function rules()
     {
@@ -62,70 +56,25 @@ class User extends Component
                             'phone' => 'nullable|regex:/^01[0-9]{9}$/',
                             'company_id' => 'required',
                             'photo' => 'nullable|image|mimes:jpeg,png,pdf|max:1024', // 1MB Max
+
                         ];
     }
 
 
-    public function updatedPhoto()
-    {
-        if (Storage::exists('app/photos/' . $this->photo->hashName())) {
-            $this->addError('photo', 'A photo with this name already exists.');
-            $this->photo = null;
-        }
-    }
-
-public function removePhoto()
-{
-    unset($this->photo);
-}
-
-public function save()
-    {
-        try{
-            if($this->photo !== null){
-                $this->photo->store('photos');
-                session()->flash('success','done successfully'); 
-            } 
-        }catch(\Exception $e){
-            session()->flash('error',$e);
-        } 
-    }
-
-public function updated($feilds)
-{
-    $this->validateOnly($feilds);
-}
 
 
 
-public function updatingSearch()
-    {
-        $this->resetPage();
-    }
-
- public function store(){
-
+ public function store(UserService $userService){
     try{
-
+        $this->check_permission($this->write_permission);
         $validatedData = $this->validate();
- 
-      $admin= Admin::create([
-     'name'=>$validatedData['name'],
-     'email'=>$validatedData['email'],
-     'company_id'=>$validatedData['company_id'],
-     'phone'=>$validatedData['phone'],
-     'password'=>Hash::make($validatedData['password']),
-        ]);
-        if($this->photo !== null){
-        $admin->addMedia($this->photo)->toMediaCollection('profile');
-        }
-        $this->reset();
-        $this->emit('closeModal');
-        session()->flash('success','done successfully');   
+        $userService->store($validatedData);
+         $this->success();
     }catch(\Exception $e){
-        session()->flash('error',$e);
-    }
- }
+        errorMessage($e);
+ }}
+
+
 
  public function edit(int $id){
  $edit= Admin::findOrFail($id);
@@ -134,6 +83,7 @@ public function updatingSearch()
   $this->admin_id= $id;
   $this->name =$edit->name;
   $this->email =$edit->email;
+  $this->phone =$edit->phone;
   $this->company_id =$edit->Company_id;
  }else{
   return redirect()->back();
@@ -142,33 +92,17 @@ public function updatingSearch()
 
  }
 
- public function closeModal()
- {
-     $this->reset('name','email','company_id','phone','password');
- }
 
- public function update(){
+
+ public function update(UserService $userService){
 
    try{
-
+    $this->check_permission($this->write_permission);
      $validatedData = $this->validate();
-    $admin= Admin::FindOrFail($this->admin_id);
-    $admin->update([
-     'name'=>$validatedData['name'],
-     'email'=>$this->email,
-     'company_id'=>$this->company_id,
-     'phone'=>$this->phone,
-     'password'=>Hash::make($validatedData['password']), 
-     ]);
-     if($this->photo !== null){
-        $admin->clearMediaCollection('profile');
-        $admin->addMedia($this->photo)->toMediaCollection('profile');
-        }
-     $this->reset();
-     $this->emit('closeModal');
-     session()->flash('success','done successfully');  
+     $userService->update( $validatedData,$this->admin_id);
+     $this->success();
     }catch(\Exception $e){
-        session()->flash('error',$e);
+          errorMessage($e);
     }
 
  }
@@ -184,25 +118,26 @@ public function updatingSearch()
 
   public function assignRoles (){ 
    try{
+    $this->check_permission($this->write_permission);
    $admin= Admin::findOrFail($this->admin_id);
-   $admin ->assignRole($this->assigned_roles);
-   session()->flash('success','done successfully'); 
-   $this->reset();
+   $admin ->syncRoles($this->assigned_roles);
+    $this->success();
    }catch(\Exception $e){
-        session()->flash('error',$e);
+    errorMessage($e);
     }
 
   }
 
    public function changeStatus (int $id){
-        $admin = Admin::findOrFail($id); 
+    $this->check_permission($this->write_permission);
+    $admin = Admin::findOrFail($id); 
     if($admin->status ==1 && $id > 1 ){
     $admin->update([
         'status' => 0
     ]);
     session()->flash('error','user now is deactive '); 
     }elseif($admin->status == 0){
-   $admin->update([
+    $admin->update([
         'status' => 1
     ]);
     session()->flash('success','user now is active '); 
@@ -212,37 +147,43 @@ public function updatingSearch()
 
   public function removeRoles (){
     try{
+        $this->check_permission($this->write_permission);
         if($this->admin_id > 1 ){
         $admin= Admin::findOrFail($this->admin_id);
         $admin ->syncRoles([]);
-        session()->flash('success','done successfully'); 
-        $this->reset();}
+        $this->success();}
         else{
-            session()->flash('error','the roles of this user are not allawed to be removed '); 
+            session()->flash('error','The roles of this admin are not allawed to be removed '); 
         }
         }catch(\Exception $e){
-             session()->flash('error',$e);
+            errorMessage($e);
          }
   }
  
- public function delete(){
 
+  
+ public function delete(){
    try {
+    $this->check_permission($this->write_permission);
+    if($this->admin_id > 1 ){
      Admin::FindOrFail($this->admin_id)->delete();
-     session()->flash('success','done successfully'); 
-   
-    }catch(\Exception $e){
-        session()->flash('error',$e);
+   $this->success();
     }
+    session()->flash('error','You are not allawed to remove this Admin '); 
+    }catch(\Exception $e){
+        errorMessage($e);
+    }
+}
+
+#[Computed]
+public function admins(){
+    return Admin::where('name', 'like', '%'.$this->search.'%')->orderBy('id', $this->sortfilter)->paginate($this->perpage);
 }
 
 
 public function render()
 {
-  $roles = Role ::all();
-   $companies = Company::all('id','name');
-    $admins = Admin:: with('roles','media')->orderBy('id', 'desc')->paginate(5);
-        $admins =Admin::where('name', 'like', '%'.$this->search.'%')->orderBy('id', 'desc')->paginate(5);
-        return view('livewire.Admin.users.user',compact('admins','roles','companies'));
+        return view('livewire.Admin.users.user');
+    
     }
 }
